Support default address better.
In the checkout UI there was no way for the user to manage their default
address. This means in general the user did not have an address attached
to their record.

Even with the UI elements there were other issues to address. Spree tends
to copy addresses around. But with the address book plugin this means
users quickly accumulated lots of addresses. To resolve I have overwritten
some spree methods to simply reference the addresses instead of copying
them around.

This commit is mainly aimed at making things better during checkout. More
work still needs to be done to make things better in the backend and in
the "My Account" area (which I don't think is part of this plugin but
which uses this plugin).

Also some dependencies were updated so the tests would work.