Delink addresses after an order is complete.
This preserve the historical data on old orders. A new copy of the
address is placed on the bill and ship address for the order as
well as the address on the shipments.

These new addresses are specifically made to NOT have an user_id
to avoid copies of the address from showing up in the user's address
book.